% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIO625HW3-package.R
\docType{package}
\name{BIO625HW3-package}
\alias{BIO625HW3}
\alias{BIO625HW3-package}
\title{BIO625HW3: Linear Regression Implementation Package}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: YU QING \email{ronnieq@umich.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
