% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QY_lm.R
\name{QY_lm}
\alias{QY_lm}
\title{Fit a linear regression model}
\usage{
QY_lm(data, formula)
}
\arguments{
\item{data}{A data frame containing the variables in the model.}

\item{formula}{A formula object describing the model (e.g., mpg ~ wt + hp).}
}
\value{
A list containing:
\item{coefficients}{The estimated coefficients.}
\item{R_squared}{The R-squared value of the model.}
\item{adj_R_squared}{The adjusted R-squared value.}
\item{residuals}{The residuals of the model.}
}
\description{
QY_lm fits a linear regression model using ordinary least squares.
}
\examples{
data(mtcars)
result <- QY_lm(mtcars, mpg ~ wt + hp)
print(result$coefficients)
}
